Added check for SMTP AUTH support
An error condition exists where the configured smtp server does not support the AUTH extension, as may be the case with a local smtp server configured to provide mail relay for services on the ctf box.  

This modification is on line 445 and will verify whether the smtp server has reported AUTH support before attempting to log in. 

Note, that this change will result in skipping the login quietly when AUTH is not supported, regardless of whether a user/pass has been configured.