Simplified code.
The two lines was using the dictionary `get()` function with a default value,
and then also testing if the same `get()` call actually returned 'something',
with an else clause with the same default value (that would newer be taken).

The type of `request.form` is `werkzeug.datastructures.ImmutableMultiDict`,
which has a rather 'special' `get()` function.  This function also takes a
'type' parameter, which (if callable) converts the value or returns the default
value if not possible.