Changed the `CHALLENGE_CLASSES` structure.
`CHALLENGE_CLASSES` now maps to a tuple of the challenge class and the
corresponding database object representing that challenge.

Thus also changing `get_chal_class()` to return the tuple.  This function used
`.get()` and tested if the returned value was `None`, in order to raise a
`KeyError`.  This has been simplified to a direct lookup in the dictionary with
`CHALLENGE_CLASSES[class_id]` as this is the exact same behaviour.

The reason to make this change is that the `CTFdStandardChallenge` now uses this
reference to the corresponding challenge database object in its `create()`
function.  This greatly simplifies subclassing the object, and now any
subclasses will work out of the box, having the correct database object
instantiated.