Challenge database object initialisation should not care about type.
The type field of a given challenge database object, should be hard coded, and
not dependent on arguments to `__init__`.  Each custom challenge database object
can easily hard code this as part of their `__init__` with a call to the static
`id` variable of its respective challenge class (e.g., `self.type =
DynamicValueChallenge.id`).

This makes it less likely that some plugins fucks up and use the wrong type, and
more importantly: it simplifies the code a bit.