Updated `Challenges` to properly handle unicode.
Turns out that sqlite3 (which nosetests use) doesn't like if we use UTF-8
encoded strings.

As per the previous commit, when we call `create()` the values will be converted
to the type of the underlying column type.  This was `str` (`db.Text` and
db.String` defaults to `convert_unicode=False`), which works fine for MySQL,
however sqlite3 has an assumption (through its `text_factory`, as far as I can
tell) about what "type" the text has.  By default in py2 this is unicode, and
thus if we try to add and commit data that is already (UTF-8) encoded, then it
errors with:

```
ProgrammingError: (sqlite3.ProgrammingError) You must not use 8-bit bytestrings
 unless you use a text_factory that can interpret 8-bit bytestrings (like
 text_factory = str).  It is highly recommended that you instead just switch
 your application to Unicode strings. [SQL: u'INSERT INTO challenges (name,
 description, max_attempts, value, category, type, hidden) VALUES (?, ?, ?, ?,
 ?, ?, ?)'] [parameters: ('\xf0\x9f\x92\xab', None, 0, 100, '\xf0\x9f\x92\xab',
 'standard', 0)]
```

This is kinda funky, as the docs only state that the `text_factory` "[...]
control what objects are returned [...]".

As stated in previous commit, [1] states that we should be using `db.Unicode` or
`db.UnicodeText`, for columns that are intended to store non-ASCII data.
Since the tests more or less only put in unicode data, this update to the
database model seems like the best way to fix the nosetests (sqlite3), without
any hacking.

[1] states that using the two new unicode column types would try to use either
`NVARCHAR`, `NCLOB` or `NTEXT`.  However MySQL doesn't support these column
types, so the prevous `VARCHAR` and `TEXT` types are still used, thus this
should not have any impact on any MySQL databases already running.

[1]: <https://docs.sqlalchemy.org/en/latest/core/type_basics.html#sqlalchemy.types.String>
[2]: <https://docs.python.org/2.7/library/sqlite3.html#sqlite3.Connection.text_factory>