Calling `util.get_config` closes all sqlalchemy sessions.
When trying to dertermine if it is "ctftime", `util.ctftime()` queries the
configuration for the start and end time using get_config.  This in tern has a
`with` statement over `app.app_context()`.

It turns out this is unneded and part of the old design, however more
importantly this closes all current sqlalchemy sessions when exiting the scope,
as part of calling the `__exit__()` function which `pop()`'s the current
context, resulting in `app.do_teardown_appcontext()` being called which loops
over all registered teardown functions (`@teardown_appcontext`).  The
flask-alchemy extension, just so happens to register such a teardown function
[1] (`shutdown_session()`).

The issue was previously not that bad, as all objects that are already fetched
from the database will still have their values.  However they are in a detached
state, and thus any changes or updates of their values is not possible.  More
importantly, if we have a polymorphic object where only some of the fields are
fetched (e.g., due to a select on the base/super class such as `Challenges`,
instead of the subclass) and the rest are marked as expired, then we can't
access these expired fields, as it would result in a new SELECT (which is
impossible due to the missing session).

[1]: <https://github.com/mitsuhiko/flask-sqlalchemy/blob/f465142ea1643b4f0fe50540780ef9a6bf2c7e53/flask_sqlalchemy/__init__.py#L801>