Support SQLALCHEMY URLS with "%" sign
Sometimes the SQLALCHEMY URL must contain a percent sign (%). One example of this is when your password contains a %. Another example is when deploying to azure mysql. The username contains an @ sign, which must be encoded as "%40" to pass to sqlalchemy as a URL. Having a percent sign in the URL breaks line 27, because the Python ConfigParser uses % as a special character for evaluating interpolated values.
In order to place the value into the Configuration, the % sign must be escaped with %%.