feat: Add option to fallback to the default theme
This is intended to make it easier to develop themes in pieces rather
than having to copy the default one over and modifying it in place, or
having to create all pages before it will work sensibly. It may not also
be easier to construct minor modifications to the core theme which
aren't easy/possible with the pages mechanism or jinja template block
overrides.