feat: Add solve count and status to challenge list
This changeset reworks the challenge list API endpoint to count the
number of solves for each challenge like the per-challenge details
endpoint already does. Both endpoints now return the `solves` key which
(for each challenge in the response for the list endpoint) which is an
integer counting the solves from visible, non-banned accounts.

If account or score visible is disabled for the current user session (or
public users), `solves` will be set to `None` rather than 0. This mostly
matches the behaviour of the `challenge/<id>/solves` endpoint  which
will not show an account's own solves if scores are not currently
visible.

This changeset also introduces a new boolean key `solved_by_me` to
challenge data JSON returned from both the challenge list and detail
endpoints. This value reflects if the challenge being described has been
solved by the currently authenticated user (or their team). It follows
the same visibility rules as `scores` and is set to `False` if the
required visibility checks fail. This could be used to avoid having to
request and parse `/{teams,users}/me/solves/` in client side code to
render solved challenges in a different colour.