feat: Allow challenge types to override visibility
The goal of this change is to support plugins which implement "stealth"
challenges which are not shown to users until they have been solved.
Such challenge types would need some way for a user to discover the
challenge IDs to make use of the traditional challenge attempt API
endpoint, or have a paired endpoint which allows the submission of flags
without specifying a challenge ID. It is expected that these details
should be implemented entirely by plugins with CTFd simply doing what it
can to consult them where reasonable.

The change introduces a property on the base challenge type class which
is used as a final visibility check prior to data for that challenge
being returned to a user. In particular, the challenge list and detail
API endpoints have been taught to check this prior to returning data.
The attempt endpoint continues to rely solely on `.attempt()` so that
challenge types which refuse to be shown to the user may still be
attempted unless the challenge type also disallows that (in which case,
what's the point?).