refac: Allow loading theme templates by prefix
This change reworks the template loaders so that we make more liberal
use of the ordered preferential loading provided by `ChoiceLoader` to
simplify the `ThemeLoader` so it only needs to be aware of a single
theme name to load templates from. This allows us to more easily
implement loading of templates from specific themes by providing an
extra `PrefixLoader` which is aware of all of the themes available to
the application.