string error Update
Since strings are not functions. In line 288, I added string_error because we are calling function () to the end of a function name. This error commonly occurs when you assign a variable called “str”, to use the str() function. Python interprets “str” as a string and you cannot use the str() function in your program