Allow teams with null passwords to create invite codes without setting their team password (#2548)
* Allow teams with null passwords to create invite codes without setting their team password

This loosens the fix implemented in https://github.com/CTFd/CTFd/pull/2485. Teams with NULL passwords can now generate invite codes that are signed with only the CTFd secret key. 

The original idea was to use both the secret key and team password to allow revocation of the invite by changing the password but this achieves the same effect as if the team sets a password, the invite generated with only the secret key will no longer work