Remove CSRF checks for safe methods
The CSRF before_request hook would occasionally modify the Flask
session. This in turn caused a new session cookie to be sent with the
response. For static content retrieved via GET requests, this behavior
is not desired, as it might cause issues with caching reverse proxies.
More specifically, we observed wrong sessions being assigned to users
via the Set-Cookie header cached in the reverse proxy, which sometimes
causes users to being logged in as a different user.

This commit therefore removes the CSRF checks for safe methods (see the
corresponding RFC) and refactors the CSRF check function to document the
CSRF exceptions and make them more explicit.